refactor: publish feature, shared with integrate
